requirements.txt
openpyxl~=3.1.5
qrcode~=8.0.0
pillow~=11.0.0
PyQt5~=5.15.11
PyQt6~=6.7.1